Rename package from 'iac_realtime' to 'iacrealtime' for consistency across files
